added resumeGame

+++ Sources/Graphics/ResumeGame/resumeGame.hpp
@@ -29,7 +29,7 @@ namespace graphic
 		graphic::IrrlichtLib *_lib;
 		irr::gui::IGUIImage *_dirigible;
 		vec2d _size;
-		size_t _count;
+		int _count;
 	};
 }
